spykes/django-gcp/requirements.txt
django==4.0.8
django-crispy-forms==1.14
django-environ==0.9.0
django-stubs==1.12.0
django-configurations[cache,database,email,search]==2.3.2
django-cors-headers==3.13.0
Pillow==9.3.0
pyyaml==6.0
uritemplate==4.1.1
psycopg2-binary==2.9.5
coverage==6.4.4
google-cloud-secret-manager==2.6.0 
django-storages[google]==1.11.1
gunicorn==20.1.0
# Brute-force attack blocking
django-axes==5.40.1